refactor(core): Move LLM provider defaults out of builder internals
agent-validation.service and agent-model-catalog.service (agents module
root) imported LLM_PROVIDER_DEFAULTS from builder/interactive/ — core
code depending on a builder-internal module. Move the file to the
agents module root so imports flow in one direction.

Part of AGENT-444 (scope item 3).

(no-changelog)
