+++ packages/cli/src/modules/agents/agent-model-catalog.service.ts
@@ -6,7 +6,7 @@ import { Service } from '@n8n/di';
 import { isModelDiscoveryProvider } from '@n8n/ai-utilities/model-discovery';
 
 import { BuilderModelLiveLookupService } from './builder/builder-model-live-lookup.service';
-import { LLM_PROVIDER_DEFAULTS } from './builder/interactive/llm-provider-defaults';
+import { LLM_PROVIDER_DEFAULTS } from './llm-provider-defaults';
 
 /** Google's models API returns ids as `models/<id>`; the AI SDK expects the bare id. */
 const GOOGLE_MODEL_ID_PREFIX = 'models/';
