+++ packages/cli/src/modules/agents/llm-provider-defaults.ts
